Properly handle filled /l regnodes and multi-char folds
Previously we were ignoring this possibility.  Suppose a regnode ends in
'S', and the next one begins with 'S", and we have compiled with /li.
Those two "SS"es in a row should match a LATIN SHARP S.  Previously they
weren't.

Multi-char folds are only valid in /l if the runtime locale is UTF-8.
But we have to construct the node so that it can handle this case.